ACR-1589: upon the first login 'items_count' is set to 0

+++ TabletCR/src/main/java/com/kaching123/tcr/InventoryHelper.java
@@ -14,7 +14,7 @@ public class InventoryHelper {
 
     private static final Uri ITEM_URI = ShopProvider.getContentUri(ItemTable.URI_CONTENT);
 
-    public static int getLimit() {
+    public static long getLimit() {
         return TcrApplication.get().getShopPref().inventoryLimit().get();
     }
